Add API endpoint to get files from connected prod, fix a lot of things

+++ controllers/actions.js
@@ -1,5 +1,8 @@
 const mysql = require("mysql");
+const path = require("path");
+const fs = require('fs');
 const checkAuth = require("../middlewares/checkAuth");
+const jwt = require("jsonwebtoken");
 const db = mysql.createConnection({
     host: process.env.DATABASE_HOST,
     user: process.env.DATABASE_USER,
@@ -7,12 +10,48 @@ const db = mysql.createConnection({
     database: process.env.DATABASE
 })
 
-exports.delete = (req, res) => {
-    const {id_delete} = req.body
-    db.query('DELETE FROM users WHERE id = ?', [id_delete], async (error, results) => {
-        if (error) {
-            console.log(error);
+exports.get = (req, res) => {
+
+
+    const token = req.cookies.access_token;
+    if(!token) {
+        res.sendFile(path.resolve("upload/default.png"), {
+            id: checkAuth.getUser(req)
+        });
+    } else {
+
+    jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
+        if (err) {
+            res.sendFile(path.resolve("upload/default.png"), {
+                id: checkAuth.getUser(req)
+            });
+        } else {
+            let number = req.query.nb;
+            if(fs.existsSync(path.resolve("upload/" + user.id + "/" + number + ".png"))) {
+
+                res.sendFile(path.resolve("upload/" + user.id + "/" + number + ".png"), {
+                    id: checkAuth.getUser(req)
+                });
+            } else {
+                res.sendFile(path.resolve("upload/default.png"), {
+                    id: checkAuth.getUser(req)
+                });
+            }
         }
     })
+    }
+}
+
+
+exports.delete = (req, res) => {
+    const {id_delete} = req.body
+
+    if(checkAuth.getUser(req) !== "none") {
+        db.query('DELETE FROM users WHERE id = ?', [id_delete], async (error, results) => {
+            if (error) {
+                console.log(error);
+            }
+        })
+    }
     res.redirect("/dashboard")
 }
